backend/requirements.txt
fastapi[standard]
supabase
sqlmodel
psycopg2
passlib
python-jose
passlib[bcrypt]
pyjwt
requests
python-dotenv
boto3>=1.28.0
python-multipart
